+++ src/component/table/table.tsx
@@ -3,6 +3,7 @@ import { StarRating } from '../star-rating/star-rating'
 import styles from './table.module.scss'
 import { TagList } from '../tag-list/tag-list'
 import { MouseEvent } from 'react'
+import { Link } from '../../bookshelf/book/link'
 
 interface Props {
     books: Array<Book>
@@ -29,7 +30,7 @@ export function BookTable({ books, onBookClick }: Props) {
                         <td>{book.metadata.title}</td>
                         <td>
                             {(book.metadata.authors || []).map((author, i) => (
-                                <div key={i}>{author}</div>
+                                <div key={i}>{author instanceof Link ? author.displayText : author}</div>
                             ))}
                         </td>
                         <td>{book.metadata.published?.getFullYear() || ''}</td>